requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
google-generativeai==0.8.3
pandas==2.2.2
Pillow==10.3.0